railway/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
supabase==2.15.2
python-multipart==0.0.20
ebooklib==0.18
beautifulsoup4==4.13.4
google-generativeai==0.8.5
httpx==0.28.1
